Separate questions in folder structure
